% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{assignCellCyclePhases}
\alias{assignCellCyclePhases}
\title{Cell cycle assignment}
\usage{
assignCellCyclePhases(object, gene.file = "~/Desktop/cell_cycle_genes.xlsx",
  do.plot = T, ...)
}
\arguments{
\item{object}{A human or mouse \code{SingleSpeciesSample}.}

\item{gene.file}{The excel sheet containing the cell cycle phase specific gene
markers. Taken from Macosko et al. 2015.}
}
\value{
A \code{data.frame} with cell cycle phase scores for every cell.
}
\description{
Assigns the cells into the 5 cell cycle phases, following the algorithm described
in Macosko et al. 2015. The table of genes is also taken from Macosko et al. 2015.
}

