% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{collapseCellsByBarcode}
\alias{collapseCellsByBarcode}
\title{Collapse cells by barcodes similarity}
\usage{
collapseCellsByBarcode(object, ...)
}
\arguments{
\item{object}{A \code{SingleSpeciesSample} object.}
}
\value{
A modified \code{SingleSpeciesSample} object.
}
\description{
Collapse cells which share 11 bases in their barcodes and only the last
one is different (usually being N, as a result of the DetectBeadSynthesisErrors
tool). The cells are marked as candidates if and only if they're classified
as belonging to the same species.
}
