% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\docType{methods}
\name{compareGeneExpressionLevels,MixedSpeciesSample-method}
\alias{compareGeneExpressionLevels,MixedSpeciesSample-method}
\title{Compare gene expression levels between two mixed species samples}
\usage{
\S4method{compareGeneExpressionLevels}{MixedSpeciesSample}(object1, object2,
  threshold1, threshold2, name1 = "sample1", name2 = "sample2")
}
\arguments{
\item{object1}{A \code{MixedSpeciesSample} object.}

\item{object2}{A \code{MixedSpeciesSample} object.}

\item{threshold1}{The threshold for the firs sample.}

\item{threshold2}{The threshold for the second sample.}
}
\value{
A plot comparing the gene expression levels by species.
}
\description{
Compare gene expression levels between two mixed species samples
}
