% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{compareGeneExpressionLevels}
\alias{compareGeneExpressionLevels}
\title{Compare gene expression levels between two single species samples}
\usage{
compareGeneExpressionLevels(object1, object2, name1 = "sample1",
  name2 = "sample2", col = "steelblue", method = "pearson", ...)
}
\arguments{
\item{object1}{A \code{SingleSpeciesSample} object.}

\item{object2}{A \code{SingleSpeciesSample} object.}

\item{name1}{The name of the first sample.}

\item{name2}{The name of the second sample.}

\item{method}{The method for computing the correlation.}
}
\value{
A plot comparing the gene expression levels.
}
\description{
Computes the averaged sums of the gene expression levels between two
samples, as well as their correlation.
}
