% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{computeGenesPerCell}
\alias{computeGenesPerCell}
\title{Compute genes per cell}
\usage{
computeGenesPerCell(object, min.reads = 2, ...)
}
\arguments{
\item{object}{A Single species sample.}
}
\value{
A \code{data.frame} with cells, gene counts and species.
}
\description{
Compute genes per cell
}

