% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{computeMitochondrialPercentage}
\alias{computeMitochondrialPercentage}
\title{Computes mitochondrial percentage}
\usage{
computeMitochondrialPercentage(object, prefix = NULL)
}
\arguments{
\item{object}{A \code{SingleSpeciesSample} object.}

\item{prefix}{The prefix of mitochondrial genes.}
}
\value{
The percentage of mitochondrial encoded RNA UMIs.
}
\description{
Computes the percentageof UMIs coming from mitochondrial encoded RNA. For human,
mouse and melanogaster samples the mitochondrial prefix is hard-coded as \code{MT-},
\code{mt-} and \code{mt:} respectively. An alternative prefix can be also provided.
}

