% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{computeTranscriptsPerCell}
\alias{computeTranscriptsPerCell}
\title{Compute transcripts per cell}
\usage{
computeTranscriptsPerCell(object, ...)
}
\arguments{
\item{object}{A Single species sample.}
}
\value{
A \code{data.frame} with cells, transcript counts and species.
}
\description{
Computes the number of transcripts (UMIs) that a cell contains.
}

