% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{extractReadStatistics}
\alias{extractReadStatistics}
\title{Extract statistics from mapped reads.}
\usage{
extractReadStatistics(object, yieldSize = 10^6, sample.name = NULL)
}
\arguments{
\item{object}{A string describing the exact location of the BAM file for the
sample under consideration.}

\item{yieldSize}{Large BAM files are processed in batches because of memory
issues. The parameter \code{yieldSize} controls the batch size. Default value: 10^6.}

\item{sample.name}{A string describing the sample.}
}
\value{
A \code{data.frame} containing the computed statistics.
}
\description{
Extract statistics from mapped reads, such as total number of reads,
uniquely mapped reaqds, percentages of intronic/intergenic etc. The function
makes use of the R package Rsamtools (part of Bioconductor).
}
