% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{findInflectionPoint}
\alias{findInflectionPoint}
\title{Computes the inflection point / cell number.}
\usage{
findInflectionPoint(object, max.cells = 10000, n = 1)
}
\arguments{
\item{object}{A vector containing the reads per cell as given by the Drop-seq pipeline (BAMTagHistogram).}

\item{max.cells}{The number of cells to consider as an upper bound and to compute the inflection point.}

\item{n}{The nubmer of inflection points to return.}
}
\value{
The cell number.
}
\description{
Computes the inflection point / cell number.
}

