% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{geneExpressionDispersion}
\alias{geneExpressionDispersion}
\title{Compute the dispersion of each gene across all cells.}
\usage{
geneExpressionDispersion(object)
}
\arguments{
\item{object}{A \code{data.frame} representing DGE.}
}
\value{
A vector of variance/mean expression values for each gene. Computation in log space.
}
\description{
Compute the dispersion of each gene across all cells.
}
