% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{geneExpressionSumUMI}
\alias{geneExpressionSumUMI}
\title{Compute the sum of transcript counts of each gene across all cells.}
\usage{
geneExpressionSumUMI(object)
}
\arguments{
\item{object}{A \code{data.frame} representing DGE.}
}
\value{
A vector of the logarithm of the sum of all UMI values for each gene.
A pseudocount is added to avoid infinities.
}
\description{
Compute the sum of transcript counts of each gene across all cells.
}

