% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{identifyDoublets}
\alias{identifyDoublets}
\title{Identify doublets}
\usage{
identifyDoublets(object)
}
\arguments{
\item{object}{A \code{DigitalGeneExpressionMatrix} object.}
}
\value{
A \code{vector} containing the assigned probabilities for each cell.
}
\description{
Subsets the DGE matrix to genes which are highly expressed and at the
same time lowly varied across all cells. Comparing the cells then probabilities
are assigned for each cell to be a doublet of the same species.
}

