% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{keepBestCells}
\alias{keepBestCells}
\title{Keep best cells according to a criterion}
\usage{
keepBestCells(object, num.cells, min.num.trans = NULL)
}
\arguments{
\item{object}{A \code{data.frame} representing the DGE.}

\item{num.cells}{The number of cells to keep, ordered by total number of UMIs.}

\item{min.num.trans}{Keep only cells with at least this number of UMIs.}
}
\value{
The processed object.
}
\description{
Keep best cells according to a criterion
}

