% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\docType{methods}
\name{listCellsToCollapse,MixedSpeciesSample-method}
\alias{listCellsToCollapse,MixedSpeciesSample-method}
\title{List cells that are candidates for collapsing.}
\usage{
\S4method{listCellsToCollapse}{MixedSpeciesSample}(object, threshold = 0.9)
}
\arguments{
\item{A}{\code{MixedSpeciesSample} object.}
}
\description{
Identify and list cells which share 11 bases in their barcodes and only the last
one is different. The cells are marked as candidates if and only if they're classified
as belonging to the same species.
}
