% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{meanGeneExpression}
\alias{meanGeneExpression}
\title{Compute the average expression of each gene over all cells}
\usage{
meanGeneExpression(object)
}
\arguments{
\item{object}{A \code{DigitalGeneExpressionMatrix} object.}
}
\value{
A vector of mean expression values for each gene.
}
\description{
Compute the average expression of each gene over all cells
}

