% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotCumulativeFractionOfReads}
\alias{plotCumulativeFractionOfReads}
\title{Plots the knee plot}
\usage{
plotCumulativeFractionOfReads(object, cutoff = 10000)
}
\arguments{
\item{object}{A \code{data.frame} read from out_readcounts.txt.gz}
}
\description{
Plots the knee plot
}

