% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotHeatmapCorrelationMatrixDGE}
\alias{plotHeatmapCorrelationMatrixDGE}
\title{Heatmap of the correlation matrix for pairs of cells.}
\usage{
plotHeatmapCorrelationMatrixDGE(object)
}
\arguments{
\item{object}{A \code{data.frame} representing the DGE matrix.}
}
\description{
Heatmap of the correlation matrix for pairs of cells.
}

