% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Histogram plot of an attribute}
\usage{
plotHistogram(object, attribute = "ATTRIBUTE MISSING")
}
\arguments{
\item{object}{A \code{data.frame} such as the outputs of \code{computeGenesPerCell}
and \code{computeTranscriptsPerCell}.}

\item{attribute}{A \code{charracter string} describing the attribute.}
}
\description{
Histogram plot of an attribute
}

