% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotMitochondrialContent}
\alias{plotMitochondrialContent}
\title{Plot mitochondrial content}
\usage{
plotMitochondrialContent(object, log_scale = True, df_names)
}
\description{
Plot mitochondrial content
}

