% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{removeCells}
\alias{removeCells}
\title{Remove cells by barcode}
\usage{
removeCells(object, cells)
}
\arguments{
\item{object}{A \code{data.frame} representing the DGE.}

\item{cells}{A character vector of cell barcodes to remove.}
}
\value{
The processed \code{data.frame} is returned.
}
\description{
Removes the provided cells from the sample.
}
