% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{removeLowQualityGenes}
\alias{removeLowQualityGenes}
\title{Filter out low quality genes}
\usage{
removeLowQualityGenes(object)
}
\arguments{
\item{object}{A \code{DigitalGeneExpressionMatrix} object.}
}
\value{
A \code{DigitalGeneExpressionMatrix} object without the
low quality genes.
}
\description{
Remove genes which are expressed in less than three cells.
}

