% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\name{splitDgeByGenesAndCellsOfSpecies}
\alias{splitDgeByGenesAndCellsOfSpecies}
\title{Split DGE by genes and cells of species}
\usage{
splitDgeByGenesAndCellsOfSpecies(object, threshold = 0.9)
}
\arguments{
\item{object}{A \code{MixedSpeciesSample} object.}

\item{threshold}{The threshold which the ratio of transcripts of one species
over the other has to surpass in order to succesfully assign a cell to a species.}
}
\value{
A list \code{data.frames} representing the DGEs for each species.
}
\description{
Split DGE by genes and cells of species
}
