% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\name{splitMixedSpeciesSampleToSingleSpecies}
\alias{splitMixedSpeciesSampleToSingleSpecies}
\title{Separate the mixed species sample into two single species samples}
\usage{
splitMixedSpeciesSampleToSingleSpecies(object, threshold)
}
\arguments{
\item{object}{A \code{MixedSpeciesSample} object.}

\item{threshold}{The threshold which the ratio of transcripts of one species
over the other has to surpass in order to succesfully assign a cell to a species.}
}
\value{
A list of two \code{SingleSpeciesSample} objects.
}
\description{
Separate the mixed species sample into two single species samples
}

