% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{collapseCellsByBarcode}
\alias{collapseCellsByBarcode}
\title{Collapse cells by barcodes similarity}
\usage{
collapseCellsByBarcode(object, ...)
}
\description{
Collapse cells which have barcodes differing by one mutation on the last base.
}

