% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\name{compareGeneExpressionLevels}
\alias{compareGeneExpressionLevels}
\title{Compare gene expression levels between two mixed species samples}
\usage{
compareGeneExpressionLevels(object1, object2, threshold1 = 0.9,
  threshold2 = 0.9)
}
\arguments{
\item{object1}{A \code{MixedSpeciesSample} object.}

\item{object2}{A \code{MixedSpeciesSample} object.}

\item{threshold1}{The threshold for the firs sample.}

\item{threshold2}{The threshold for the second sample.}
}
\value{
A plot comparing the gene expression levels by species
}
\description{
Compare gene expression levels between two mixed species samples
}

