% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{computeCorrelationSingleCellsVersusBulk}
\alias{computeCorrelationSingleCellsVersusBulk}
\title{Compute gene expression levels correlation of a single cell sample against bulk data}
\usage{
computeCorrelationSingleCellsVersusBulk(single.cells, bulk.data,
  method = "pearson")
}
\arguments{
\item{single.cells}{The single cell object.}

\item{bulk.data}{The bulk data formed as a \code{data.frame} with genes as
rownames and RPKM values as the first column.}

\item{method}{The method to compute the correlation.}
}
\value{
A list of length 2. First item is the correlation and second is the
\code{data.frame}.
}
\description{
Compute gene expression levels correlation of a single cell sample against bulk data
}

