% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{computeGenesPerCell}
\alias{computeGenesPerCell}
\title{Compute genes per cell}
\usage{
computeGenesPerCell(object, min.umis = 1, ...)
}
\arguments{
\item{object}{A \code{SingleSpeciesSample}.}

\item{min.umis}{The minimum number of UMIs to consider the gene as expressed (default is 1).}
}
\value{
A \code{data.frame} with cells, gene counts and species.
}
\description{
Computes the number of genes that a cell expresses.
}
