% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSpeciesSample.R
\name{computeMitochondrialPercentage}
\alias{computeMitochondrialPercentage}
\title{Computes percentage of UMIs coming from mitochondrial encoded RNA.}
\usage{
computeMitochondrialPercentage(object, prefix = NULL)
}
\arguments{
\item{object}{A \code{SingleSpeciesSample} object.}

\item{prefix}{The prefix of mitochondrial genes.}
}
\value{
The percentage of mitochondrial encoded RNA UMIs.
}
\description{
Computes percentage of UMIs coming from mitochondrial encoded RNA.
}

