% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{findInflectionPoint}
\alias{findInflectionPoint}
\title{Compute the inflection point / cell number.}
\usage{
findInflectionPoint(object, max.cells = 10000, n = 1)
}
\arguments{
\item{object}{A vector containing the reads per cell as given by the Drop-seq
pipeline (BAMTagHistogram).}

\item{max.cells}{The number of cells to consider as an upper bound and to
compute the inflection point (default is 10000).}

\item{n}{The nubmer of inflection points to return (default is 1).}
}
\value{
The cell number.
}
\description{
Compute an inflection point (a.k.a. knee point, elbow of curve) of curve.
Can be used to compute the number of STAMPS that are present in the data
(see Drop-seq computational cookbook for further details).
}
