% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{geneExpressionMean}
\alias{geneExpressionMean}
\title{Compute the average expression of each gene across all cells.}
\usage{
geneExpressionMean(object)
}
\arguments{
\item{object}{A \code{DigitalGeneExpressionMatrix} object.}
}
\value{
A vector of mean expression values for each gene. Computation in log space.
}
\description{
Compute the average expression of each gene across all cells.
}

