% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotCellTypes}
\alias{plotCellTypes}
\title{Plot separation of species}
\usage{
plotCellTypes(object)
}
\arguments{
\item{object}{A \code{data.frame} produced by the \code{classifyCellsAndDoublets}
function.}
}
\description{
Scatter plot of all cells by transcripts of species (a.k.a. barnyard plot). Species
and doublets are separated by different colors.
}
