% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingMethods.R
\name{plotMitochondrialContent}
\alias{plotMitochondrialContent}
\title{Plot mitochondrial content}
\usage{
plotMitochondrialContent(object, log_scale = TRUE,
  sample_names = paste0("sample", 1:length(object)))
}
\arguments{
\item{object}{A list of mitochondrial content percentages, as computed by the
\code{ComputeMitochondrialContent} function.}

\item{log_scale}{Should the y-axis be in the log-scale?}

\item{sample_names}{The sample names}
}
\description{
Violin plots of mitochondrial content of one or more samples.
}

