% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{removeLowQualityCells}
\alias{removeLowQualityCells}
\title{Filter out low quality cells}
\usage{
removeLowQualityCells(object, ...)
}
\arguments{
\item{object}{A \code{DigitalGeneExpressionMatrix} object.}
}
\value{
A \code{DigitalGeneExpressionMatrix} object without the
low quality cells.
}
\description{
Remove cells which express less than 2000 genes.
}

