% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DGEMatrix.R
\name{removeLowQualityGenes}
\alias{removeLowQualityGenes}
\title{Filter out low quality genes}
\usage{
removeLowQualityGenes(object, n = 3)
}
\arguments{
\item{object}{A \code{data.frame} representing DGE.}

\item{n}{The minimum number of cells.}
}
\value{
A \code{data.frame} without the low quality genes.
}
\description{
Remove genes which are expressed in less than a minimum number of cells.
}

