% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedSpeciesSample.R
\name{splitDgeByGenesOfSpecies}
\alias{splitDgeByGenesOfSpecies}
\title{Split DGE by genes of species}
\usage{
splitDgeByGenesOfSpecies(object)
}
\arguments{
\item{object}{A \code{MixedSpeciesSample} object.}
}
\value{
A list of \code{data.frames} corresponding to the genes of the two species.
}
\description{
Split DGE by genes of species
}

